Refactor fitDataset into separate file: training_dataset.ts
